requirements.txt
requests
PyYAML
python-dotenv
pytest
pytest-mock
responses
